+++ file-content-extraction/tests/test_operations.py
@@ -6,13 +6,13 @@
 """
 Copyright start
 MIT License
-Copyright (c) 2024 Fortinet Inc
+Copyright (c) 2025 Fortinet Inc
 Copyright end
 """
 
 import pytest
 from testframework.conftest import valid_configuration, invalid_configuration, valid_configuration_with_token,\
-    connector_id, connector_details, info_json, params_json
+    connector_id, connector_details, info_json, params_json, validate_and_cleanup, initial_setup
 from testframework.helpers.test_helpers import run_health_check_success, run_invalid_config_test, run_success_test,\
     run_output_schema_validation, run_invalid_param_test, set_report_metadata
     
@@ -21,7 +21,7 @@ from testframework.helpers.test_helpers import run_health_check_success, run_inv
 def test_check_health_success(valid_configuration, connector_details):
     set_report_metadata(connector_details, "Health Check", "Verify with valid Configuration")
     result = run_health_check_success(valid_configuration, connector_details)
-    assert result.get('status') == 'Available'
+    assert result.get('status', '').lower() == 'available'
     
 
 @pytest.mark.extract_text
@@ -31,13 +31,6 @@ def test_extract_text_success(cache, valid_configuration_with_token, connector_d
                                    action_params=params_json['extract_text']):
         assert result.get('status') == "Success"
 
-
-@pytest.mark.extract_text
-def test_validate_extract_text_output_schema(cache, valid_configuration_with_token, connector_details,
-                                                 info_json, params_json):
-    set_report_metadata(connector_details, "Extract Text", "Validate Output Schema")
-    run_output_schema_validation(cache, 'extract_text', info_json, params_json['extract_text'])
-    
 
 @pytest.mark.extract_text
 def test_extract_text_invalid_file_iri(valid_configuration_with_token, connector_details, params_json):
@@ -56,13 +49,6 @@ def test_extract_indicators_success(cache, valid_configuration_with_token, conne
 
 
 @pytest.mark.extract_indicators
-def test_validate_extract_indicators_output_schema(cache, valid_configuration_with_token, connector_details,
-                                                 info_json, params_json):
-    set_report_metadata(connector_details, "Extract Artifacts", "Validate Output Schema")
-    run_output_schema_validation(cache, 'extract_indicators', info_json, params_json['extract_indicators'])
-    
-
-@pytest.mark.extract_indicators
 def test_extract_indicators_invalid_file_iri(valid_configuration_with_token, connector_details, params_json):
     set_report_metadata(connector_details, "Extract Artifacts", "Verify with invalid File IRI/Path")
     result = run_invalid_param_test(connector_details, operation_name='extract_indicators', param_name='file_iri',
@@ -77,13 +63,6 @@ def test_extract_indicators_from_file_success(cache, valid_configuration_with_to
                                    action_params=params_json['extract_indicators_from_file']):
         assert result.get('status') == "Success"
 
-
-@pytest.mark.extract_indicators_from_file
-def test_validate_extract_indicators_from_file_output_schema(cache, valid_configuration_with_token, connector_details,
-                                                 info_json, params_json):
-    set_report_metadata(connector_details, "Extract Artifacts From File", "Validate Output Schema")
-    run_output_schema_validation(cache, 'extract_indicators_from_file', info_json, params_json['extract_indicators_from_file'])
-    
 
 @pytest.mark.extract_indicators_from_file
 def test_extract_indicators_from_file_invalid_file_iri(valid_configuration_with_token, connector_details, params_json):
@@ -101,9 +80,9 @@ def test_get_backend_config_success(cache, valid_configuration_with_token, conne
         assert result.get('status') == "Success"
 
 
-@pytest.mark.get_backend_config
-def test_validate_get_backend_config_output_schema(cache, valid_configuration_with_token, connector_details,
-                                                 info_json, params_json):
-    set_report_metadata(connector_details, "Get Backend Config", "Validate Output Schema")
-    run_output_schema_validation(cache, 'get_backend_config', info_json, params_json['get_backend_config'])
-    
+@pytest.mark.create_xslx_file_from_json_data
+def test_create_xslx_file_from_json_data_success(cache, valid_configuration_with_token, connector_details, params_json):
+    set_report_metadata(connector_details, "Create XSLX file as attachment from JSON Data", "Verify with valid Input Parameters")
+    for result in run_success_test(cache, connector_details, operation_name='create_xslx_file_from_json_data',
+                                   action_params=params_json['create_xslx_file_from_json_data']):
+        assert result.get('status') == "Success"