+++ file-content-extraction/release_notes.md
@@ -1,3 +1,6 @@
 #### What's Improved
 
-Added new action `Extract Artifacts From File`
+## Added following new actions and playbooks: 
+
+- Extract Artifacts Extended
+- Create XSLX File as Attachment From JSON Data